pyproject.toml
[project]
name = "mse-mlops"
version = "0.1.0"
description = "DINOv3 fine-tuning setup for HAM10000 skin lesion classification."
authors = [{ name = "mse-mlops contributors" }]
readme = "README.md"
keywords = ["python"]
requires-python = ">=3.12,<3.13"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "datetime>=6.0",
    "imageio>=2.37.3",
    "mlflow>=2.22.1",
    "numpy>=2.4.2",
    "pandas>=2.3.2",
    "pandas-stubs~=2.3.3",
    "pillow>=12.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]

[project.urls]
Homepage = "https://7ben18.github.io/mse-mlops/"
Repository = "https://github.com/7ben18/mse-mlops"
Documentation = "https://7ben18.github.io/mse-mlops/"

[dependency-groups]
api = [
    "fastapi",
    "pydantic",
    "python-multipart",
    "uvicorn[standard]",
]
ui = [
    "openpyxl",
    "requests",
    "streamlit",
]
dev = [
    "pytest>=7.2.0",
    "ruff>=0.6.9",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "mkdocs-jupyter>=0.24.0",
]

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
